refactor: move screen-specific components to subfolder
