+++ src/screens/Gurbani/Line.spec.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { render } from '@testing-library/react-native'
 
-import Languages from '../lib/languages'
-import transliterators from '../lib/transliterators'
+import Languages from '../../lib/languages'
+import transliterators from '../../lib/transliterators'
 
 import Line, { LineProps } from './Line'
